+++ fdbclient/Tenant.cpp
@@ -134,32 +134,6 @@ TenantMapEntry::TenantMapEntry(int64_t id, TenantName tenantName, Optional<Tenan
   : tenantName(tenantName), tenantGroup(tenantGroup) {
 	setId(id);
 }
-TenantMapEntry::TenantMapEntry(MetaclusterTenantMapEntry metaclusterEntry)
-  : tenantName(metaclusterEntry.tenantName), tenantLockState(metaclusterEntry.tenantLockState),
-    tenantGroup(metaclusterEntry.tenantGroup), configurationSequenceNum(metaclusterEntry.configurationSequenceNum) {
-	setId(metaclusterEntry.id);
-}
-
-MetaclusterTenantMapEntry::MetaclusterTenantMapEntry() {}
-MetaclusterTenantMapEntry::MetaclusterTenantMapEntry(int64_t id,
-                                                     TenantName tenantName,
-                                                     TenantAPI::TenantState tenantState)
-  : tenantName(tenantName), tenantState(tenantState) {
-	setId(id);
-}
-MetaclusterTenantMapEntry::MetaclusterTenantMapEntry(int64_t id,
-                                                     TenantName tenantName,
-                                                     TenantAPI::TenantState tenantState,
-                                                     Optional<TenantGroupName> tenantGroup)
-  : tenantName(tenantName), tenantState(tenantState), tenantGroup(tenantGroup) {
-	setId(id);
-}
-
-void MetaclusterTenantMapEntry::setId(int64_t id) {
-	ASSERT(id >= 0);
-	this->id = id;
-	prefix = TenantAPI::idToPrefix(id);
-}
 
 void TenantMapEntry::setId(int64_t id) {
 	ASSERT(id >= 0);
@@ -202,47 +176,6 @@ json_spirit::mObject TenantGroupEntry::toJson() const {
 	}
 
 	return tenantGroupEntry;
-}
-
-std::string MetaclusterTenantMapEntry::toJson() const {
-	json_spirit::mObject tenantEntry;
-	tenantEntry["id"] = id;
-
-	tenantEntry["name"] = binaryToJson(tenantName);
-	tenantEntry["prefix"] = binaryToJson(prefix);
-
-	tenantEntry["tenant_state"] = TenantAPI::tenantStateToString(tenantState);
-	tenantEntry["assigned_cluster"] = binaryToJson(assignedCluster);
-
-	if (tenantGroup.present()) {
-		tenantEntry["tenant_group"] = binaryToJson(tenantGroup.get());
-	}
-
-	tenantEntry["lock_state"] = TenantAPI::tenantLockStateToString(tenantLockState);
-	if (tenantState == TenantAPI::TenantState::RENAMING) {
-		ASSERT(renameDestination.present());
-		tenantEntry["rename_destination"] = binaryToJson(renameDestination.get());
-	} else if (tenantState == TenantAPI::TenantState::ERROR) {
-		tenantEntry["error"] = error;
-	}
-
-	return json_spirit::write_string(json_spirit::mValue(tenantEntry));
-}
-
-bool MetaclusterTenantMapEntry::matchesConfiguration(MetaclusterTenantMapEntry const& other) const {
-	return tenantName == other.tenantName && tenantLockState == other.tenantLockState &&
-	       tenantGroup == other.tenantGroup && assignedCluster == other.assignedCluster;
-}
-
-void MetaclusterTenantMapEntry::configure(Standalone<StringRef> parameter, Optional<Value> value) {
-	if (parameter == "tenant_group"_sr) {
-		tenantGroup = value;
-	} else if (parameter == "assigned_cluster"_sr && value.present()) {
-		assignedCluster = value.get();
-	} else {
-		TraceEvent(SevWarnAlways, "UnknownTenantConfigurationParameter").detail("Parameter", parameter);
-		throw invalid_tenant_configuration();
-	}
 }
 
 TenantMetadataSpecification<TenantMapEntry>& TenantMetadata::instance() {
