+++ fdbserver/workloads/TenantManagementConcurrencyWorkload.actor.cpp
@@ -170,17 +170,16 @@ struct TenantManagementConcurrencyWorkload : TestWorkload {
 
 	ACTOR static Future<Void> createTenant(TenantManagementConcurrencyWorkload* self) {
 		state TenantName tenant = self->chooseTenantName();
-		state TenantMapEntry entry;
-		state MetaclusterTenantMapEntry mEntry;
-		entry.tenantName = mEntry.tenantName = tenant;
-		entry.tenantGroup = mEntry.tenantGroup = self->chooseTenantGroup();
+		state MetaclusterTenantMapEntry entry;
+		entry.tenantName = tenant;
+		entry.tenantGroup = self->chooseTenantGroup();
 
 		try {
 			loop {
 				Future<Void> createFuture =
 				    self->useMetacluster
-				        ? MetaclusterAPI::createTenant(self->mvDb, mEntry, AssignClusterAutomatically::True)
-				        : success(TenantAPI::createTenant(self->dataDb.getReference(), tenant, entry));
+				        ? MetaclusterAPI::createTenant(self->mvDb, entry, AssignClusterAutomatically::True)
+				        : success(TenantAPI::createTenant(self->dataDb.getReference(), tenant, TenantMapEntry(entry)));
 				Optional<Void> result = wait(timeout(createFuture, 30));
 				if (result.present()) {
 					break;
